Refactor async email background thread
In preparation for adding more background task processing, refactor
the async email processing into a more generic background thread
plus a function in grouper.email_util that specifically sends the
queued mail messages.

There are no functionality changes in this diff, just moving code
around and changing a class method to a regular function.
